add rowsAffected and _version handling to deleteTask
also enables the deleteTask for all delete requests.
but the alias check still needs to be done.

+++ sql/src/main/java/io/crate/metadata/doc/DocSysColumns.java
@@ -17,7 +17,7 @@ public class DocSysColumns {
 
     private static ImmutableMap<ColumnIdent, DataType> columnIdents = ImmutableMap.<ColumnIdent, DataType>builder()
             .put(new ColumnIdent("_id"), DataType.STRING)
-            .put(new ColumnIdent("_version"), DataType.INTEGER)
+            .put(new ColumnIdent("_version"), DataType.LONG)
             .put(new ColumnIdent("_score"), DataType.DOUBLE)
             .build();
